feat: add reorder JSON accept middleware
We need this for MCP routes so the 'authenticate' middleware returns a JSON 401, even if the JSON isn't the first preference in the 'Accept' header.

This happens in VS Code for example where its Accept header asks for text-event-stream first, then application/json, but it still expects a JSON 401 response or it breaks.